Complete: implemented scry
Works. There's a bug I haven't figured out when scrying with no cards in
deck (line 565) but that's an ok edge case for this sort of project.
Relies on DECKREMOVE always getting topmost copy of card--can we change
the guarantee to be that? since that's actually what it does right now.